cli: refactor setup_plugin_{args,options}
- Move funcs from `streamlink_cli.main` to `streamlink_cli.argparser`
- Pass `Streamlink` session and `argparse.Namespace` objects as params
- Use the session's `UserInputRequester` rather than referencing
  the `console` object and require it to exist in the session options
- Update function references in `_parser` module (docs helper utility)
- Fix typing error in `options.Argument.dest`
- Refactor tests and move to `test_argparser`
