+++ src/streamlink_cli/_parser.py
@@ -4,8 +4,7 @@ when building the documentation, man page or command-line shell completions.
 """
 
 from streamlink.session import Streamlink
-from streamlink_cli.argparser import build_parser
-from streamlink_cli.main import setup_plugin_args
+from streamlink_cli.argparser import build_parser, setup_plugin_args
 
 
 def get_parser():
